+++ lib/build.gradle
@@ -22,7 +22,6 @@ dependencies {
     implementation "com.squareup.okhttp3:okhttp:$versions.okhttp"
 
     testImplementation 'org.assertj:assertj-core:3.19.0'
-    testImplementation "com.google.truth:truth:1.1.3"
     testImplementation "org.junit.jupiter:junit-jupiter-api:${versions.jupiter}"
     testImplementation "org.junit.jupiter:junit-jupiter-params:${versions.jupiter}"
     testRuntime "org.junit.jupiter:junit-jupiter-engine:${versions.jupiter}"
